+++ R/learner_survivalmodels_surv_akritas.R
@@ -55,8 +55,6 @@ LearnerSurvAkritas = R6Class("LearnerSurvAkritas",
 
   private = list(
     .train = function(task) {
-      assert_python_packages(c("torch", "pycox", "numpy"), python_version = "3.10")
-
       pv = self$param_set$get_values(tags = "train")
       invoke(
         survivalmodels::akritas,
@@ -68,8 +66,6 @@ LearnerSurvAkritas = R6Class("LearnerSurvAkritas",
     },
 
     .predict = function(task) {
-      assert_python_packages(c("torch", "pycox", "numpy"), python_version = "3.10")
-
       pv = self$param_set$get_values(tags = "predict")
       newdata = ordered_features(task, self)