+++ NEWS.md
@@ -1,6 +1,9 @@
 # mlr3extralearners (development version)
 
-* fix: remove `discrete` parameter from `surv.parametric`, so that it is impossible to return `distr6::VectorDistribution` survival predictions (softly deprecated in `mlr3proba@v0.8.1`)
+* feat: Added learner `LearnerCompRisksRandomForestSRC`
+* feat: Added `selected_features` property to RandomForestSRC learners (prediction doesn't work if `vars.used = 'all.trees'`)
+* fix: Removed parameters from RandomForestSRC learners that weren't used + optimized tests
+* fix: Remove `discrete` parameter from `surv.parametric`, so that it is impossible to return `distr6::VectorDistribution` survival predictions (softly deprecated in `mlr3proba@v0.8.1`)
 * feat: Added learner `LearnerSurvBlockForest`.
 * feat: Added learner `LearnerClassifFastai`.
 * feat: Added learners `LearnerRegrPenalized` and `LearnerClassifPenalized`.